+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## Beta 1.6 (tba)
+
+### Added
+
+- Added "Temporarily Disable Physics Toggling" button in the settings. If your vehicle gets stuck in the air, you can temporarily disable physics toggling. Physics toggling will be enabled back next time you restart the game
+- Added current date to mod report
+
+### Changes
+
+- Changed the initialization delay to 2 seconds from 1
+
+### Bug Fixes
+
+- Fixed battery charger cables connecting themselves, after leaving and getting back to home
+- Changed how GT grille is checked on initial loading. GT grille should not disappear anymore (at least, I hope so)
+
 ## Beta 1.5 (10.01.2020)
 
 ### Added